fix CustomListDao_Impl.java UNIQUE constraint failed: CustomList.id, CustomList.position, CustomList.idCanto

+++ app/src/main/java/it/cammino/risuscito/CantiEucarestiaFragment.kt
@@ -1,7 +1,6 @@
 package it.cammino.risuscito
 
 import android.content.Intent
-import android.database.sqlite.SQLiteConstraintException
 import android.os.Bundle
 import android.os.SystemClock
 import android.preference.PreferenceManager
@@ -359,80 +358,56 @@ class CantiEucarestiaFragment : Fragment() {
 
     private fun scambioCanto(v: View, position: Int) {
         val idNew = Integer.valueOf((v.findViewById<View>(R.id.text_id_canto_card) as TextView).text.toString())
-        val timestampNew = (v.findViewById<View>(R.id.text_timestamp) as TextView).text.toString()
-        //        Log.i(getClass().toString(), "positionNew: " + position);
-        //        Log.i(getClass().toString(), "idNew: " + idNew);
-        //        Log.i(getClass().toString(), "timestampNew: " + timestampNew);
-        //        Log.i(getClass().toString(), "posizioneDaCanc: " + posizioneDaCanc);
-        //        Log.i(getClass().toString(), "idDaCanc: " + idDaCanc);
-        //        Log.i(getClass().toString(), "timestampDaCanc: " + timestampDaCanc);
         if (idNew != idDaCanc || posizioneDaCanc != position) {
-            val roomDb = RisuscitoDatabase.getInstance(context!!)
-            roomDb.beginTransaction()
-
-            try {
-                val positionToDelete = CustomList()
-                positionToDelete.id = 2
-                positionToDelete.position = position
-                positionToDelete.idCanto = idNew
-                val mDao = RisuscitoDatabase.getInstance(context!!).customListDao()
+            val mDao = RisuscitoDatabase.getInstance(context!!).customListDao()
+            if (mDao.checkExistsPosition(2, position, idDaCanc) > 0
+                    || mDao.checkExistsPosition(2, posizioneDaCanc, idNew) > 0) {
+                Snackbar.make(rootView!!, R.string.present_yet, Snackbar.LENGTH_SHORT).show()
+            } else {
+                val positionToDelete = mDao.getPositionSpecific(2, position, idNew)
                 mDao.deletePosition(positionToDelete)
-
-                mDao.updatePositionNoTimestamp(idNew, 2, posizioneDaCanc, idDaCanc)
 
                 val positionToInsert = CustomList()
                 positionToInsert.id = 2
                 positionToInsert.position = position
                 positionToInsert.idCanto = idDaCanc
-                positionToInsert.timestamp = Date(java.lang.Long.parseLong(timestampNew))
+                positionToInsert.timestamp = positionToDelete.timestamp
+
+                mDao.updatePositionNoTimestamp(idNew, 2, posizioneDaCanc, idDaCanc)
                 mDao.insertPosition(positionToInsert)
 
-                roomDb.setTransactionSuccessful()
                 Snackbar.make(
                         activity!!.findViewById(R.id.main_content),
                         R.string.switch_done,
                         Snackbar.LENGTH_SHORT)
                         .show()
-            } catch (e: SQLiteConstraintException) {
-                Snackbar.make(rootView!!, R.string.present_yet, Snackbar.LENGTH_SHORT).show()
-            } finally {
-                roomDb.endTransaction()
             }
-
         } else {
             Snackbar.make(rootView!!, R.string.switch_impossible, Snackbar.LENGTH_SHORT).show()
         }
     }
 
     private fun scambioConVuoto(position: Int) {
-        val roomDb = RisuscitoDatabase.getInstance(context!!)
-        roomDb.beginTransaction()
+        val mDao = RisuscitoDatabase.getInstance(context!!).customListDao()
 
-        try {
-            val positionToDelete = CustomList()
-            positionToDelete.id = 2
-            positionToDelete.position = posizioneDaCanc
-            positionToDelete.idCanto = idDaCanc
-            val mDao = RisuscitoDatabase.getInstance(context!!).customListDao()
+        if (mDao.checkExistsPosition(2, position, idDaCanc) > 0)
+            Snackbar.make(rootView!!, R.string.present_yet, Snackbar.LENGTH_SHORT).show()
+        else {
+            val positionToDelete = mDao.getPositionSpecific(2, posizioneDaCanc, idDaCanc)
             mDao.deletePosition(positionToDelete)
 
             val positionToInsert = CustomList()
             positionToInsert.id = 2
             positionToInsert.position = position
             positionToInsert.idCanto = idDaCanc
-            positionToInsert.timestamp = Date(java.lang.Long.parseLong(timestampDaCanc!!))
+            positionToInsert.timestamp = positionToDelete.timestamp
             mDao.insertPosition(positionToInsert)
 
-            roomDb.setTransactionSuccessful()
             Snackbar.make(
                     activity!!.findViewById(R.id.main_content),
                     R.string.switch_done,
                     Snackbar.LENGTH_SHORT)
                     .show()
-        } catch (e: SQLiteConstraintException) {
-            Snackbar.make(rootView!!, R.string.present_yet, Snackbar.LENGTH_SHORT).show()
-        } finally {
-            roomDb.endTransaction()
         }
     }
